requirements.txt
accelerate==0.23.0
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
azure-core==1.26.4
blinker==1.6.2
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
confection==0.0.4
cymem==2.0.7
dataclasses-json==0.5.7
decorator==5.1.1
diffusers==0.21.2
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
fugashi==1.2.1
ginza==5.1.2
greenlet==2.0.2
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.8.0
ipadic==1.0.0
itsdangerous==2.1.2
ja-ginza==5.1.2
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.175
langcodes==3.3.0
lazy_loader==0.2
librosa==0.10.0.post2
lit==16.0.5
llvmlite==0.40.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mecab-python3==1.0.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numba==0.57.0
numexpr==2.8.4
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.1
pathy==0.10.1
Pillow==9.5.0
plac==1.3.5
pooch==1.6.0
preshed==3.0.8
psutil==5.9.5
pycparser==2.21
pydantic==1.10.7
pyre-extensions==0.0.29
PySoundFile==0.9.0.post1
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
rhoknp==1.3.0
safetensors==0.3.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
soundfile==0.12.1
soxr==0.3.5
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.15
srsly==2.4.6
SudachiDict-core==20230110
SudachiPy==0.6.7
sympy==1.12
tenacity==8.2.2
thinc==8.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.2
triton==2.0.0
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
unidic==1.1.0
unidic-lite==1.0.8
urllib3==2.0.2
wasabi==0.10.1
wcwidth==0.2.6
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.17.0
